requirements.txt
langchain>=1.1.0
langgraph>=1.0.4
pinecone>=7.3.0
streamlit>=1.51.0
openai>=2.8.1
anthropic>=0.75.0
python-dotenv>=1.2.1
requests>=2.28.0
numpy>=1.24.0
perplexityai>=0.17.0
langchain-pinecone>=0.2.13
langchain-perplexity>=1.1.0